@@ -1,8 +1,8 @@
 source 'https://rubygems.org'
-ruby '2.3.4'
+ruby '2.5.7'
 
-gem 'rails', '4.2.5.1'
-gem 'pg', '~> 0.15'
+gem 'rails', '~> 5.2.4'
+gem 'pg', '~> 0.21'
 gem 'sass-rails', '~> 5.0'
 gem 'uglifier', '>= 1.3.0'
 gem 'therubyracer'
@@ -10,14 +10,16 @@ gem 'therubyracer'
 gem 'dotenv'
 gem 'jquery-rails'
 gem 'turbolinks'
-gem 'sdoc', '~> 0.4.0', group: :doc
+gem 'bootsnap', require: false
+gem 'coffee-rails'
+gem 'sdoc', '~> 1.0.0', group: :doc
 
 gem 'rack-cache'
-gem 'bcrypt', '~> 3.1.7'
-gem 'puma', '~> 3.2'
+gem 'bcrypt', '~> 3.1.12'
+gem 'puma', '~> 4.3'
 gem 'rails_12factor'
 
-gem 'simple_form', '~> 3.2'
+gem 'simple_form', '~> 5.0'
 
 gem 'jwt'
 gem 'premailer-rails'
@@ -26,25 +28,26 @@ gem 'httparty'
 gem 'diffy'
 gem 'kramdown'
 
-# only used for alerting SQS.
-gem 'aws-sdk', '~> 2.0'
+gem 'aws-sdk-sqs', '~> 1.30'
 
 group :development, :test do
   gem 'byebug'
-  gem 'rspec-rails', '~> 3.0'
+  gem 'rspec-rails', '~> 3.8'
   gem 'guard-rspec', require: false
-  gem 'factory_girl_rails', '~> 4.0'
+  gem 'factory_bot_rails', '~> 6.1'
   gem 'database_cleaner'
   gem 'sinatra'
 end
 
 group :test do
   gem 'webmock'
+  gem 'rails-controller-testing'
+  gem 'rspec_junit_formatter'
   gem 'codeclimate-test-reporter', require: nil
 end
 
 group :development do
   gem 'letter_opener'
-  gem 'web-console', '~> 2.0'
+  gem 'web-console', '~> 3.7'
   gem 'spring'
 end
