@@ -1,4 +1,4 @@
-class User < ActiveRecord::Base
+class User < ApplicationRecord
   validates :email, length: { minimum: 3 }, uniqueness: { case_sensitive: false }
   validate :email_domain_is_approved, on: [ :create, :update ]
   has_many :pages
@@ -35,7 +35,7 @@ class User < ActiveRecord::Base
 
   def send_notification(change)
     puts "user#send_notification #{self.email}"
-    ChangeMailer.page(change: change, user: self).deliver_later
+    ChangeMailer.page(change: change, user: self).deliver_now
   end
 
   def email_domain_is_approved
