@@ -1,4 +1,4 @@
-class CreateAppSettings < ActiveRecord::Migration
+class CreateAppSettings < ActiveRecord::Migration[4.2]
   def change
     create_table :app_settings do |t|
       t.string :key
