@@ -1,4 +1,4 @@
-class CreateSqsIntegrations < ActiveRecord::Migration
+class CreateSqsIntegrations < ActiveRecord::Migration[4.2]
   def change
     create_table :sqs_integrations do |t|
       t.text :queue_url
