@@ -1,7 +1,7 @@
-class AddIsAdminToUser < ActiveRecord::Migration
+class AddIsAdminToUser < ActiveRecord::Migration[4.2]
   def change
     add_column :users, :is_admin, :boolean
-    
+
     # Set new `is_admin` field for users listed in the `ADMIN_EMAILS`
     # environment variable, if it exists
     emails = ENV['ADMIN_EMAILS'].to_s.split(',')
@@ -13,16 +13,16 @@ class AddIsAdminToUser < ActiveRecord::Migration
       user.is_admin = true
       user.save
     end
-    
+
     # The below only should apply to people hosting Klaxon themselves rather
     # than on Heroku, and only in certain situations.
     if emails.empty?
       puts <<~EOF
         Your ADMIN_EMAILS environment variable was empty.
-        
+
         If you want to be able to delete users from within the Klaxon
         Web interface, set up at least one admin user.
-        
+
         The easiest way to do this is to run the "users:create_admin"
         Rake task with the ADMIN_EMAILS environment variable set.
       EOF
