@@ -1,4 +1,4 @@
-FactoryGirl.define do
+FactoryBot.define do
 
   factory :page_snapshot do
     page { create(:page) }
@@ -9,12 +9,12 @@ FactoryGirl.define do
   end
 
   factory :slack_integration do
-    channel "#klaxon"
-    webhook_url "http://test-webhook.com/test-webhook"
+    channel { "#klaxon" }
+    webhook_url { "http://test-webhook.com/test-webhook" }
   end
 
   factory :sqs_integration do
-    queue_url "https://sqs.us-east-1.amazonaws.com/1234567890/klaxon-sqs-q-test"
+    queue_url { "https://sqs.us-east-1.amazonaws.com/1234567890/klaxon-sqs-q-test" }
   end
 
   factory :change do
@@ -27,13 +27,13 @@ FactoryGirl.define do
   end
 
   factory :page do
-    name "nyt homepage"
-    url "http://www.nytimes.com/"
-    css_selector "h2.story-heading"
+    name { "nyt homepage" }
+    url { "http://www.nytimes.com/" }
+    css_selector { "h2.story-heading" }
 
     trait :with_snapshots do
       transient do
-        snapshot_count 2
+        snapshot_count { 2 }
       end
       after(:create) do |page, evaluator|
         create_list(:page_snapshot, evaluator.snapshot_count, page: page)
