@@ -14,7 +14,7 @@ RSpec.describe SlackIntegration, type: :model do
 
     payload = slack.send_notification(change)
     expect(payload[:icon_url]).to start_with  'http'
-    expect(payload[:icon_url]).to include '/assets/klaxon-logo'
+    expect(payload[:icon_url]).to include '/images/klaxon-logo'
     expect(payload[:channel]).to eq slack.channel
     expect(payload[:username]).to include "Klaxon"
     expect(payload[:text]).to include page.name
