@@ -8,11 +8,10 @@ RSpec.describe Change, type: :request do
 
     # login
     @user = User.where(email: 'test@test.com').first_or_create
-    get(token_session_path, token: LoginToken.create(user: @user))
+    get(token_session_path, params: { token: LoginToken.create(user: @user) })
   end
 
   it "can create multiple snapshots and send an email based on the last pair" do
-
     stub_request(:any, /faketimeserver.com/).to_rack(FakeTimeServer)
 
     page = Page.create!(url: "http://faketimeserver.com/now", user: @user, css_selector: 'body')
@@ -27,12 +26,6 @@ RSpec.describe Change, type: :request do
 
     Change.check
 
-    assert_enqueued_jobs 1 # even though we had three snapshots, expect just one email
-
-    ## TODO: WIP!
-
+    expect(ActionMailer::Base.deliveries.length).to eq 1 # even though we had three snapshots, expect just one email
   end
-
-
-
 end
